mailsender/requirements.txt
amqp==2.6.1
astroid==1.6.6
backports.functools-lru-cache==1.6.4
billiard==3.6.4.0
celery==4.4.7
configparser==4.0.2
contextlib2==0.6.0.post1
DAWG-Python==0.7.2
Django==1.11.29
django-ckeditor==5.9.0
django-js-asset==1.2.3
docopt==0.6.2
enum34==1.1.10
flake8==3.9.2
functools32==3.2.3.post2
futures==3.4.0
importlib-metadata==2.1.3
isort==4.3.21
kombu==4.6.11
lazy-object-proxy==1.6.0
mccabe==0.6.1
pathlib2==2.3.7.post1
Pillow==6.2.2
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytz==2022.7.1
redis==3.5.3
scandir==1.10.0
singledispatch==3.7.0
six==1.16.0
typing==3.10.0.0
vine==1.3.0
wrapt==1.14.1
zipp==1.2.0
